Squashed commit of the following:
commit ca5d0033b8f455ee156731189749044304f2ea22
Author: kaustubh <kaustubhkj@gmail.com>
Date:   Sat Sep 6 15:19:19 2025 +0530

    tests fixed

commit f24d78206f58cc1412780d0c381c6517c51c3e42
Author: kaustubh <kaustubhkj@gmail.com>
Date:   Sat Sep 6 13:49:43 2025 +0530

    still fix tests

commit adaeb15fa721a7769c3acd3e7eacb93acb19e8b8
Author: kaustubh <kaustubhkj@gmail.com>
Date:   Sat Sep 6 13:44:30 2025 +0530

    need to fix tests

commit 424a3375830a52d911225f83af0dd07fb4bff9d3
Author: kaustubh <kaustubhkj@gmail.com>
Date:   Fri Sep 5 16:56:20 2025 +0530

    reflection given normal

commit 8d998134a846e5854af066f46c92c0bd4e180d68
Author: kaustubh <kaustubhkj@gmail.com>
Date:   Wed Sep 3 15:01:40 2025 +0530

    normal calculation wrt sphere

+++ playground/ray_sphere_intersect.go
@@ -9,10 +9,10 @@ import (
 
 func Process() {
 	ray_origin := coordinates.CreatePoint(0, 0, -5)
-	wall_z := float32(10.0)
-	wall_size := float32(7.0)
+	wall_z := 10.0
+	wall_size := 7.0
 
-	canvas_pixels := float32(100)
+	canvas_pixels := 100.0
 	pixel_size := wall_size / canvas_pixels
 	half := wall_size / 2
 
@@ -22,10 +22,10 @@ func Process() {
 	sph := rays.NewSphere(coordinates.CreatePoint(0, 0, 0), 1)
 
 	for y := 0; y < my_canvas.GetHeight(); y++ {
-		world_y := half - pixel_size*float32(y)
+		world_y := half - pixel_size*float64(y)
 
 		for x := 0; x < my_canvas.GetWidth(); x++ {
-			world_x := -half + pixel_size*float32(x)
+			world_x := -half + pixel_size*float64(x)
 
 			position := coordinates.CreatePoint(world_x, world_y, wall_z)
